+++ app/src/main/kotlin/com/revolgenx/anilib/common/data/store/StoreModule.kt
@@ -6,7 +6,7 @@ import com.revolgenx.anilib.common.data.store.theme.ThemeDataStore
 import com.revolgenx.anilib.common.data.store.theme.WidgetThemeDataStore
 import com.revolgenx.anilib.list.data.store.MediaListEntryEventStore
 import com.revolgenx.anilib.notification.data.store.NotificationDataStore
-import com.revolgenx.anilib.widget.viewmodel.AiringWidgetResource
+import com.revolgenx.anilib.widget.ui.AiringScheduleWidgetResource
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.qualifier.named
 import org.koin.core.scope.Scope
@@ -21,7 +21,7 @@ val storeModules = module {
     single { AppPreferencesDataStore(get()) }
     single { MediaListEntryEventStore() }
     single { NotificationDataStore(get()) }
-    single { AiringWidgetResource(get(), get()) }
+    single { AiringScheduleWidgetResource(get(), get()) }
 }
 
 fun Scope.animeListFilterDataStore() = androidContext().animeListFilterDataStore